refactor(favorite): favorite repository in memory separated into files
Files created to add favorite, fetch favorite by id and remove favorite
