+++ src/infra/favorite/implementation/repositories/in-memory/fetch-favorite-by-id-repository-in-memory/fetch-favorite-by-id-repository-in-memory.ts
@@ -1,0 +1,24 @@
+import {
+  Favorite,
+  IFetchFavoriteByIdRepository,
+} from '../../../../../../domain/favorite';
+import { database } from '../../../../../database-in-memory';
+import { FavoriteAdapter } from '../../../../adapters';
+
+export class FetchFavoriteByIdRepositoryInMemory
+  implements IFetchFavoriteByIdRepository
+{
+  async fetchFavoriteById(id: number): Promise<Favorite | undefined> {
+    const favorite = database.favorites.find(
+      (favorite: any) => favorite.favorite_id === id
+    );
+
+    if (!favorite) {
+      return undefined;
+    }
+
+    return FavoriteAdapter.fromJson({
+      ...favorite,
+    });
+  }
+}
