+++ src/infra/favorite/implementation/repositories/in-memory/index.ts
@@ -1,0 +1,3 @@
+export * from './add-favorite-repository-in-memory/add-favorite-repository-in-memory';
+export * from './fetch-favorite-by-id-repository-in-memory/fetch-favorite-by-id-repository-in-memory';
+export * from './remove-favorite-repository-in-memory/remove-favorite-repository-in-memory';
