+++ src/infra/favorite/implementation/repositories/in-memory/remove-favorite-repository-in-memory/remove-favorite-repository-in-memory.ts
@@ -1,0 +1,14 @@
+import { IRemoveFavoriteRepository } from '../../../../../../domain/favorite';
+import { database } from '../../../../../database-in-memory';
+
+export class RemoveFavoriteRepositoryInMemory
+  implements IRemoveFavoriteRepository
+{
+  async removeFavorite(id: number): Promise<void> {
+    const index = database.favorites.findIndex(
+      (favorite: any) => favorite.favorite_id === id
+    );
+
+    database.favorites.splice(index, 1);
+  }
+}
