+++ src/main/favorite/factories/use-cases/add-favorite-use-case-factory.ts
@@ -1,12 +1,12 @@
 import { AddFavoriteUseCase } from '../../../../domain/favorite';
-import { FavoriteRepositoryInMemory } from '../../../../infra/favorite/implementation/repositories/in-memory/favorite-repository-in-memory';
+import { AddFavoriteRepositoryInMemory } from '../../../../infra/favorite/implementation/repositories/in-memory';
 import { HeroRepositoryInMemory } from '../../../../infra/hero/repositories/in-memory/hero-repository-in-memory';
 import { UserRepositoryInMemory } from '../../../../infra/user/implementation/repositories/in-memory/user-repository-in-memory';
 
 export const makeAddFavoriteUseCase = () => {
   const fetchUserByIdRepository = new UserRepositoryInMemory();
   const fetchHeroByIdRepository = new HeroRepositoryInMemory();
-  const addFavoriteRepository = new FavoriteRepositoryInMemory();
+  const addFavoriteRepository = new AddFavoriteRepositoryInMemory();
   return new AddFavoriteUseCase(
     fetchUserByIdRepository,
     fetchHeroByIdRepository,
