+++ src/main/favorite/factories/use-cases/fetch-favorite-by-id-use-case-factory.ts
@@ -1,7 +1,8 @@
 import { FetchFavoriteByIdUseCase } from '../../../../domain/favorite';
-import { FavoriteRepositoryInMemory } from '../../../../infra/favorite/implementation/repositories/in-memory/favorite-repository-in-memory';
+import { FetchFavoriteByIdRepositoryInMemory } from '../../../../infra/favorite/implementation/repositories/in-memory';
 
 export const makeFetchFavoriteByIdUseCase = () => {
-  const fetchFavoritesByIdRepository = new FavoriteRepositoryInMemory();
+  const fetchFavoritesByIdRepository =
+    new FetchFavoriteByIdRepositoryInMemory();
   return new FetchFavoriteByIdUseCase(fetchFavoritesByIdRepository);
 };
